backend/requirements.txt
# Flask and extensions
flask==1.1.4
werkzeug==1.0.1
flask-cors==3.0.10
flask-restx==0.4.0
flask-sqlalchemy>=2.5.1,<3.0.0
flask-migrate==2.7.0
flask-script==2.0.6

# Database and Cloud SQL
cloud-sql-python-connector==1.2.0
pg8000==1.29.8
SQLAlchemy>=1.4.46,<2.0.0
psycopg2-binary==2.9.9

# Google Cloud
google-cloud-translate>=3.0.0
google-generativeai>=0.3.2

# JWT for authentication
PyJWT==2.8.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
markupsafe==2.0.1
Jinja2==2.11.3

# Testing
pytest==7.4.0
pytest-cov==4.1.0

#ML configs
scikit-learn
pandas
numpy

#Highlights
google-cloud-storage
ffmpeg-python
moviepy ==1.0.3
google-cloud-aiplatform